fix: Only include project LICENSE and NOTICE in Spark Client Jar
Previously the build code would add LICENSE and NOTICE files
from other shaded dependencies, some of which could potentially
 shade the project's LICENSE or NOTICE or cause confusion as to
which file was correct.

To fix this a post-processing task is added which removes all
potentially confusing files and manually adds the project
LICENSE and NOTICE files back to the jar.