feat: Support customizing S3 endpoints (#1913)
No functional change.

Introduce a dedicated interface for `StsClient` suppliers and implement it using a pool of cached clients.

All client are "thin" and share the same `SdkHttpClient`. The latter is closed when the server shuts down.

This is a step towards supporting non-AWS S3 storage (#1530).
For this reason the STS endpoint is present in new interfaces, but is not used yet.