Update CatalogEntity::Builder to set default CatalogType as INTERNAL (#1998)
Encountered the issue while adding additional validations to `ExternalCatalog`. The `CatalogEntity::Builder` checks if the `Catalog::type` is set to `INTERNAL`, if not it defaults to `EXTERNAL`. However this is the opposite of the behavior defined in polaris-management-service.yml where the default is set to `INTERNAL`. This change only affects tests because in other cases the catalog entity is generated from the REST request.  

Testing:
Updated CatalogEntityTest to ensure that the default is set to `INTERNAL`.