Extract ResolverFactory from PolarisEntityManager
`PolarisEntityManager.prepareResolver` was the only place that used the
`EntityCache`.
By introducing an application-scoped `ResolverFactory` the creation of
an `EntityCache` is decided only once in `QuarkusProducers`, which will become
the only place where `MetaStoreManagerFactory.getOrCreateEntityCache` gets called.

Note that the little left-over functionality of `PolarisEntityManager` can
be split out to more dedicated interfaces in a follow-up most likely.