Simplify bootstrapServiceAndCreatePolarisPrincipalForRealm (#2172)
this is a small follow-up to 5faa3712cb8f0ae8140d096565734432d5391eb1
because the same pattern existed for this method.

note that we do some minor additional "formatting" changes to minimize
the diff between the two files (as they were originally copy pasted).
this could lead to having a common base class in the future.