Remove overloads of BasePersistence.listEntities
`BasePersistence.listEntities` and `TransactionalPersistence.listEntitiesInCurrentTxn`
are afaict only overloaded to make the `entityFilter` and `transformer`
parameters "optional" for some callers.
the most central method still has to have support for both of them.

using default methods in the interface would enable this more clearly
but as it turns out, hardly any callers are utilizing those overloads
and thus if we simply remove the overloads we have less code and things
become clearer across the board.

we also add a `EntityNameLookupRecord.fromEntity` factory method to
replace the overloaded constructor.