Standardize logging libraries in tests (#2268)
This change enforces the following test logging patterns:

- Non-Quarkus modules use Logback Classic, configured via logback-test.xml
- Quarkus modules use JBoss Logging Manager, configured in Quarkus configuration files.

This change also introduces a workaround for the "duplicate log messages" issues with Gradle + JBoss Logging Manager. See this issue for context:

https://github.com/quarkusio/quarkus/issues/22844

The workaround implemented in this PR is very similar to the one proposed in this comment:

https://github.com/quarkusio/quarkus/issues/22844#issuecomment-1824720923

Note: it's not entirely possible imho to suppress the following message on the console:

```
The Agroal dependency is present but no JDBC datasources have been defined.
```

This is because:

1. The message happens during augmentation phase, not during tests
2. And it suffers from the "duplicate message" issue with (it's actually Gradle that prints those messages).