Rework PolarisMetaStoreManager.listEntities
the central `BasePersistence.listEntities` method supports pagination,
filtering loaded entities via a Predicate and transforming the loaded
entities to a more specific type.

`PolarisMetaStoreManager.listEntities` only exposed a limited subset of
that functionality i.e. the retrieval of a `ListEntitiesResult`.

almost all callers then had to post-process the lookup records of
`ListEntitiesResult` and called `PolarisMetaStoreManager.loadEntity`
on the individual items to transform and filter them.

this is bad for the following reasons:
- suboptimal performance as we run N+1 queries to basically load every
  entity twice from the persistence backend
- suffering from race-conditions when entities get dropped between the
  `listEntities` and `loadEntity` call
- a lot of repeated code in all the callers (of which only some are
  dealing with the race-condition by filtering out `null` values)

the fix is to simply let `PolarisMetaStoreManager.listEntities` offer
all capabilities of the underlying `BasePersistence.listEntities`
method.
as it turns out `ListEntitiesResult` and `PolarisEntity.NameAndId`
become completely obsolete.

note that we remove `testCatalogNotReturnedWhenDeletedAfterListBeforeGet`
from `ManagementServiceTest` because the simulated race-condition
scenario can not happen any longer.