Remove `CallContext.copy()`
The function is only needed to create a "safe clone" of `PolarisCallContext` in `TaskExecutorImpl`, which doesn't have references to CDI request scoped beans.

To still let `TaskExecutorImpl` making "safe clones", a functionality to get (fresh) instances of `RealmContext` is required. To enable this, the `RealmContextResolver` has been enhanced with "`RealmContext` lookups" by realm-ID. That in turn led to splitting the HTTP/REST-to-realm-context resolution into two parts: HTTP/REST-to-realm-ID and realm-ID-to-context.

This change is part of the effort to allow tasks run run in different JVMs.