Add feature config to allow dropping views without purging (#2369)
* Add feature config to allow dropping views without purging

With tables, the client can decide whether to purge the table
on drop or not. However, Polaris Servers used to unconditionally
perform the purge on dropping a view.

After #1619 that behaviour effectively prevents dropping views
if the admin user does not set `DROP_WITH_PURGE_ENABLED`. The
latter, though, is not currently advisable per #1617.

This change introduces a new feature configuration
(`PURGE_VIEWS_ON_DROP`) that allows the admin user to instruct
Polaris servers to drop views without purging to achieve
operational parity with tables.

Fixes #2367

* review: rename to PURGE_VIEW_METADATA_ON_DROP

* review: re-fix description