Pass related entity to PolarisStorageIntegrationProvider
PolarisStorageIntegrationProvider used to operate on a
`PolarisStorageConfigurationInfo` instance extracted from
the related entity by the caller.

This change delegated this config extraction to the implementations
of the `PolarisStorageIntegrationProvider` interface.

The main purpose of this change is to allow implementations to
have more flexibility by being able to inspect the whole entity
that contains storage info.

The default `PolarisStorageIntegrationProvider` implementation
works as before. There is no behaviour change in Polaris from
the clients' perspective.

As a side effect, `IntegrationPersistence` and `TransactionalPersistence`
interfaces also had to be changed to propagate related entities to
`PolarisStorageIntegrationProvider`.