Remove hadoop-common from polaris-runtime-service
since https://github.com/apache/iceberg/commit/9bc6853938533c9c1dfac497229c8f4c30672458
passing a hadoop Configuration to `CatalogUtil.loadCatalog` is optional.

since `hadoop-common` brings in a lot of dependencies (and thus
potential vulnerabilities) we should avoid depending on it in our core
modules if not strictly required.

note that we were using the transitive `commons-collections` dependency
in a spot that we have to adjust.