Restore original Request ID functionality
This PR enacts the action items decided in the following ML thread:

https://lists.apache.org/thread/4gmlcn84o7n8d12qnpl7grfbm4zypk7b

The action items are:

1. Restore the original functionality for request IDs, change the
default header name back to x-request-id
2. Remove RequestIdGenerator and related functionality.
3. Update PolarisEvent, expose request ID if available, expose OTel
context if available.
4. Update events table SQL schema: insert request ID if available,
insert OTel context if available.

Furthermore, this PR also fixes #2913 and adds a small integration test for the JDBC events sink: `InMemoryBufferEventListenerIntegrationTest`.

Finally, it modifies the Helm chart to reflect the configuration changes.