NoSQL: Node IDs - API, SPI + general implementation (#2728)
* NoSQL: Node IDs - API, SPI + general implementation

This PR provides a mechanism to assign a Polaris-cluster-wide unique node-ID to each Polaris instance, which is then used when generating Polaris-cluster-wide unique Snowflake-IDs.

The change is fundamental for the NoSQL work, but also demanded for the existing relational JDBC persistence.

Does not include any persistence specific implementation.

* NoSQL: Fail node-management-impl init after timeout

Also move the expensive part to a `@PostConstruct` to not block CDI entirely from initializing.