Remove runtime-common and defaults from service deps
The main idea is to avoid exposing RDS (added in #2650) as a primary dependency of
runtime-service since it is conceptually only a dependency on leaf modules (like
runtime-server, admin and tests).

This is expected to simplify dependency management in downstream projects since
runtime-service has a reduced set of transitive dependencies now.