Update release guide to reference the proposed vote e-mail
Now that #3150 has been merged, a VOTE e-mail is pre-generated by the
third workflow so that there is less room for error.  The release guide
has been updated to reflect this.

There is no pre-generated Incubator vote e-mail as part of the third
workflow.  This is acceptable in that eventually, after graduation, that
would become unnecessary.  But also that it could cause confusion as
there would be two proposed e-mail bodies in the third workflow.
Recommendation is to let the release manager created the Incubator vote
e-mail manually.